power: Temporarily disable EcWakeupHelper
CL:329948 added a new feature to powerd that needs to call into ectool
to control the EC keyboard wake angle. Unfortunately, I only just now
Discovered that ectool is only available in test images, and that
deploying it on production images is not desired.

This patch temporarily disables the EcWakeupHelper that was introduced
by that patch until we can implement the appropriate kernel interface to
make this work again. It's supposed to be a short-term band-aid to avoid
spamming crash reports and intentionally not removing any code yet,
since it's not completely clear how the new interface will look and
which parts we'll still need yet.

BUG=chrome-os-partner:50993,chromium:594037
TEST=Emerged with tests on Minnie and Falco.

Change-Id: I8480445df76f460250c818ed4af5ba48269e22ea
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/332260
Reviewed-by: Dan Erat <derat@chromium.org>