power: Launch sync when the power button is pressed
Shutting down by holding the power button for a few seconds is
considered a supported way of turning off the system, and so should
not cause any filesystem issues. We may not be syncing early enough
in the shutdown process, so we should sync when the user first
presses the power button.

BUG=chromium:591078
TEST=press power button, look for "Launching sync" in logs

Change-Id: Ifa142f947b57db016ef0ce7ff237b1a115d9f31e
Reviewed-on: https://chromium-review.googlesource.com/332231
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>