power: Reenable EcWakeupHelper via sysfs access
This patch reenables the EcWakeupHelper class that was disabled in
CL:332260 and implements the EC keyboard wake angle control through
this interface. Since it changed from 3.14 to 3.18 kernels and we want
to support older devices as well, we need to implement support and
detection for two slightly different interfaces.

Also remove the old powerd_setuid_helper code that calls ectool since it
is no longer needed.

CQ-DEPEND=CL:333075
BUG=chromium:594037
TEST=Deployed on Minnie, confirmed that keyboard wakeup behavior was as
expected. Built with unit tests for Falco. Deployed on Oak and confirmed
that kb_wake property on the EC was written correctly.

Change-Id: I4c263feec6f35c3811ff4ec3149a18a59fd61257
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/333038
Commit-Ready: Dan Shi <dshi@google.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>