login: Add --oobe-bootstrapping-slave flag to ChromeBit (veyron_mickey).
This change is meant for M49. The bootstrapping feature (a ChromeOS device is capable of being enrolled into a domain automatically
via Bluetooth) should have been enabled by default in M49. However, currently we accidently keep the feature behind a command switch.
So we decided to add a command switch to ChromeBit only and merge it into M49. After that, we'll enable the feature by default in M50
and later.

BUG=None
TEST=Run image on device and check the command switch in chrome://version page.

Change-Id: I4414534d85b53dd1365661aa2eb9781aef7d66a9
Reviewed-on: https://chromium-review.googlesource.com/333668
Commit-Ready: Xiaoqian Dai <xdai@chromium.org>
Tested-by: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>