libcontainer: turn on user namespacing
This change will enable user namespacing for the android container.
It has a couple of marked hacks, that will be removed as soon as the
dependent issues are resolved.
Mounting noexec outside the container is removed.  It doesn't provide
any real security as root can always run things from /proc/pid/root of
the container anyways.  The new container user isn't able to change the
mount flags, so it is impossible now in addition to being ineffective.

BUG=b/24976046
TEST=boot on minnie and cyan, run some apps

Change-Id: I3a123ef06104385f10b1307a83d331de6a6b265b
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/338710