quipper: Output 40-byte perfized build ID strings
The perf build IDs are stored in PerfReader with length that could differ
from the standard 40 bytes. However, when outputting a filename-to-build
ID map, the output build IDs should be formatted to a 40-byte length.

Previously, the build ID was stored as a 40-byte array and also output as
such. After the change [https://chromium-review.googlesource.com/#/c/323513/],
the output build IDs could be truncated to a shorter length, since they are
now stored internally as a variable-length array.

By padding the build IDs to 40 bytes in GetFilenamesToBuildIDs(), this patch
restores the behavior of that function from before the aforementioned CL.

BUG=chromium:568870
TEST=unit tests pass

Change-Id: Ic434c03b6d2bddbf9f1e8ec7e75331957306c6d8
Reviewed-on: https://chromium-review.googlesource.com/335465
Commit-Ready: Simon Que <sque@chromium.org>
Tested-by: Simon Que <sque@chromium.org>
Reviewed-by: David Sharp <dhsharp@chromium.org>