crash: Collect console-ramoops for hardware watchdog resets
Many of our boards contain a hardware watchdog that reboots the machine
if it hasn't been petted for a while. This serves as the last line of
defense against the most unrecoverable of hangs. Right now we only
detect these events as an unclean shutdown, but we do not differentiate
them from other such events (e.g. battery ran dry in S3) and do not
collect any debugging information.

This patch adds support to crash-reporter to detect hardware watchdog
resets and upload the console-ramoops file in a crash report in that
case (subject to the usual consent and sanitization rules, of course).
console-ramoops contains a ring buffer with the most recent dmesg lines
before the last reboot, similar to the normal pstore crash dumps we
collect for real kernel crashes. Since the kernel cannot respond to a
hardware watchdog timeout this dump will not contain the usual backtrace
and register dump, but the last few log messages before the hang can
still often give a useful clue about the problem. We aggregate it
through a "signature" computed from the last log line which is the most
likely one related to the problem.

Since crash-reporter needs to read /var/log/eventlog.txt for this, we
change the userfeedback job that creates this file to a 'task' which
starts on 'starting crash-reporter-late'. This makes upstart wait for it
to finish before actually starting the crash-reporter-late job.

CQ-DEPEND=CL:335264
BUG=chromium:595531
TEST=Tried both 'cat > /dev/watchdog' and 'echo c > /proc/sysrq-trigger'
on Oak, observed how the expected crash dumps got written to
/var/spool/crash. Ran unit tests.

Change-Id: I89fa18917926fbb458733b5c5ca5204fe219ad37
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/334910
Reviewed-by: Mike Frysinger <vapier@chromium.org>