crash: Split --init into --init and --boot_collect
The crash-reporter --init job is used for two distinct things: to
initialize some sysfs settings that cause the system to collect crashes
when they happen, and to run certain one-shot per-boot collection tasks.
We want to ensure that the former happens as soon as possible (to be
able to record crashes as early as possible), while the latter may
include longer-running operations and dependencies on other tasks but is
not very time-critical. Therefore, this patch splits the two into two
separate upstart tasks, so that crash-reporter could get moved further
up the boot chain while crash-boot-collect can accumulate more
dependencies without fear of holding up anything critical.

Also remove the --nounclean_check option, because this patch makes it
unnecessary for the autotest that seems to have been its only user.

CQ-DEPEND=CL:335294
BUG=chromium:595531
TEST=Booted on Oak with --verbose command line and manually confirmed
that both jobs ran in the expected order. Ran 'test_that e:logging_.*'.

Change-Id: I000db5ff41bc170708e3a8699a6d911231b0efe8
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/335262