Reland "crash: Move crash-reporter before boot-services"
User crash handling should be enabled as early as possible to make sure
we can detect crashes early in the boot process. Since most of the
crash_reporter initialization code has now been split out into
crash-boot-collect, the crash-reporter job really just consists of a few
lightweight writes to sysfs. It should be okay to run it in a hook
before boot-services to allow crash handling for almost all user-space
processes (although we still need to wait for syslog).

One complication is that the crash_reporter binary still always sets up
its DBus connections, even if the current execution mode doesn't require
DBus for anything. This can be fixed by moving SetUpDbus() out of
Initialize() and instead make it a function that has to be called
on-demand before every DBus access.

This was reverted previously due to race conditions in the syslog
startup scripts.  Those should be resolved as of CL:639130.

BUG=chromium:199893
TEST=unit tests and logging_.*

Change-Id: Ic254a40d4e5491a5a8809ca482bc4e4008df4997
Reviewed-on: https://chromium-review.googlesource.com/657438
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>