crash: Collect Java crash logs from ARC
This CL extends the ARC collector to create reports for Java crashes in
ARC. A Java crash log is a text file containing headers with metadata,
followed by content dependent on the crash type. The content for most
crash types is a Java stack trace.

Java crash logs are collected in ARC, and sent across the ARC bridge to
Chrome, which forwards them to crash_reporter. The crash type is passed
with a command-line switch, and the log is read from standard input. The
collector writes the raw log to a .log file, as well as exception info
to a .info file if applicable.

BUG=b:26673021
BUG=b:27503472
TEST=Trigger non-native ARC crash using dev tools. Verify the exception
     info, log, and metadata in ~chronos/crash.

Change-Id: I60a802e0ef96a79ed2d75fb51723582808b2f1c9
Reviewed-on: https://chromium-review.googlesource.com/334282
Commit-Ready: Dominik Laskowski <domlaskowski@google.com>
Tested-by: Dominik Laskowski <domlaskowski@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>