debugd: Close the FD after GetBigFeedbackLogs() to signal readers on the other end of the pipe.
Initially I wrote the Chrome client side code to wait until it gets the response from D-Bus
to start reading the file contents. Now I want to use a pipe reader, and for it to work properly,
the server must close the FD in order to signal the pipe reader to finish reading the contents.
Otherwise, the pipe reader will wait forever.

I also added an empty table |big_feedback_logs| that will be filled later with entries of huge
log files needed to be part of the feedback reports.

BUG=b:27210832
TEST=cros_workon_make --board=peach_pit chromeos-base/debugd --test

Change-Id: Id654f3907c466ddb421e12fe13da527f54ceb1f1
Reviewed-on: https://chromium-review.googlesource.com/336302
Commit-Ready: Ahmed Fakhry <afakhry@chromium.org>
Tested-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>