crash: dump more data for i915 debug
We met a couple of GPU hang issues where the FR is very low(about 3%).
We have checked a lot of logs, however, the information is not enough to
root cause the problem. Besides, we can not reproduce this problem
at our end. So, to get sufficient debug information, we need to dump
more data by default when GPU hang. In this patch, we dump i915/parameter,
i915_capabilities and i915_wa_registers.

BUG=chrome-os-partner:50629
BRANCH=TOT
TEST=/sbin/crash_reporter --udev=KERNEL=card0:SUBSYSTEM=drm:ACTION=change
    check log under folder /var/spool/crash/change_card0_drm.x.y.x.log

Change-Id: I4fccad781e4a3c211c154df6e94f3b387ed88d50
Signed-off-by: Zhuo-hao Lee <zhuo-hao.lee@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/336760
Reviewed-by: Dan Erat <derat@chromium.org>