shill: Remove progressive scan
This code is dead and adds a bunch of extra complexity to the
scanning logic. This removes progressive scan from the available
scan types and gets rid of ScanSession.

This helps determine correctness of extensions to the scanning
code, since we only use one pathway (which goes through wpa_supplicant)
out of the two that currently exist.

BUG=chromium:602716
TEST=unittests

Change-Id: Ic9d3217be16fa425216351615d919be051abf301