installer: install MLO and u-boot.img if available
In the PostInstall phase for legacy u-boot bios (like Beaglebone) also
copy /boot/MLO and /boot/u-boot.img if available to EFI system
partition. This is needed because changing settings in u-boot like
loadaddr or fdaddr are not reflected in the EFI partition which can
render the system unbootable.

BUG=None
TEST=update u-boot for beablebone, build a new image and use
     update_engine_client to update, check that new MLO and u-boot.img
     are copied to EFI partition.

Change-Id: I98348275f47116392d13cffcc07dcb79ec0b67c5
Reviewed-on: https://chromium-review.googlesource.com/340481
Commit-Ready: Victor Dodon <printesoi@chromium.org>
Tested-by: Victor Dodon <printesoi@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>