shill: don't process proxy packet after read error
IOInputHandler::OnFileCanReadWithoutBlocking calls the error handler
on a read error, but it also calls the input handler afterwards.
We should really change the behavior of the IOInputHandler,
but I don't know if some callers rely on the current behavior, so
I am fixing this only for the HTTP proxy case.  Also clean up the
confusion between ssize_t, size_t, and int (everything should be
ssize_t here).

BUG=crosbug.com/p/52689
TEST=none

Change-Id: I433c45ee5e3e8a4539a49ec38aa9562b68bfad2e