libcontainer: Make config struct const in container
config isn't changed and can be const.  This allows the caller to use a
const struct for container config and leaves ownership of the config
struct with the caller.  It is now the callers responsibility to free
the config struct as well.  There aren't any callers committed yet, so
nothing is affected.

BUG=none
TEST=boot container on minnie.

Change-Id: I0d62982b0715323975a0ab6562b4bdbfee55a93a
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/341077