libcontainer: Allow relative source when mounting externally
Only mounts made inside minijail were checked for being relative to the
runtime root.  Allow mounts made outside to be relative as well.  Thus,
if a mount needs to be made externally, it can still be specified
relative to the runtime root path of the container.
This allows the bind mount of /dev/pts/ptmx to /dev/ptmx to be made
outside and have its selinux context set correctly.

BUG=none
TEST=run container on minnie with root-relative, external mount.

Change-Id: I8b4703aae20b40dbad021c35e01b6fdbfa2175f6
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/341585
Reviewed-by: Brian Norris <briannorris@chromium.org>