shill: Initate VPN reconnections when the default physical service is Online
The current logic tries to reconnect the VPN when the default physical
service changes, and is in any of the connected states.  This means
that it will try to reconnect even if a portal was detected and shill
is waiting for the user to authenticate to the portal.  This is not
desirable because the device does not have internet access in Portal
state.

Change it so that it waits until the default service is Online before
attempting a reconnection.

Bug: None
BUG=chromium:598781
TEST=`FEATURES=test emerge-link shill`

(cherry picked from commit 8e3ac1b9da21aac8c0e97049e0a0c8326f32e3ba)

Change-Id: I42d6a90e16e5b3a3834fba6e0e249b0d45226e03
Reviewed-on: https://chromium-review.googlesource.com/345706
Commit-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>