crash: Excise core_collector dependencies
In order to collect native ARC crashes on x86_64, core_collector must be
built as an i686 executable. However, multilib support for libbrillo and
libchrome is not implemented, so core_collector cannot depend on them.
This CL removes the dependencies, and updates the ARC collector to use
the new command-line interface.

BUG=b:28304295
TEST=cat core | core_collector  # Still works on ARM.

Change-Id: I62794ce8f71bc8ddec97f4f8c3e367d8f64c0b9e
Reviewed-on: https://chromium-review.googlesource.com/341781
Commit-Ready: Dominik Laskowski <domlaskowski@google.com>
Tested-by: Dominik Laskowski <domlaskowski@google.com>
Reviewed-by: Luis Hector Chavez <lhchavez@google.com>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>