installer: Look for kernel partition in root device first.
To prevent reading mmcblk0rmpb or mmcblk0rootX (these are rather
expensive operation because we have to switch eMMC hardware partition),
check first if the kernel partition the machine booted from is part of
the root device. If not, try all the blocks are before.

While there, add "${...}" around variable names.

BUG=chromium:609661
TEST=Check with sh -x that the flow is correct.
Tamper rootdev and check that alternate patch is taken.

Change-Id: I7884441f5e06e7366bb698f5d8f7c5f31c5b1e68
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/342844
Reviewed-by: Mike Frysinger <vapier@chromium.org>