login: VPD update implemented as part of policy store.
The policy store flow in Chrome OS part starts in
SessionManagerDBusAdaptor::StorePolicy() method.

The old implementation was using a DBusMethodCompletion object
to pass to SessionManagerImpl and to PolicyService, so that
whenever an error occurs or the process is finished, the completion
object is used to notify Chrome through DBus call about the flow result.

The object is destroyed when the function call is finished and it
checks in destructor that the completion object was used. See
DBusMethodCompletion::HandleResult from SessionManagerDBusAdaptor
and its destructor.

Chrome considers that enrollment is done when it receives the response
through DBus call from that object.

Since VPD update is running in a separate process, we need to know when
the process finishes. So, the VpdProcessImpl was converted to be a
JobManagerInterface and we are notified when the process is finished
through HandleExit method.

Since the VPD update is the last step in the flow of policy store, the
VpdProcessImpl object will run the completion when it finishes. This
will send the D-Bus response back to Chrome.

The functionality to update VPD was moved to DevicePolicyService class
and additional cc file was added for dbus_error namespace so that the
constants inside act can be passed between methods.

BUG=chromium:438234
TEST=manually tested only lumpy device and unit tests.
CQ-DEPEND=CL:336357, CL:344511

Change-Id: Ib6e884969532515f20a1060cdd000831a467a86d
Reviewed-on: https://chromium-review.googlesource.com/336438
Commit-Ready: Igor Covganet <igorcov@chromium.org>
Tested-by: Igor Covganet <igorcov@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Thiemo Nagel <tnagel@chromium.org>