boot-splash: add dependency on udev-trigger-early
Which makes sure graphics and tty devices required for frecon
to work are initialized. Otherwise frecon would segfault because
of missing /dev/ptmx or start using display device before it is
fully initialized. This adds a trigger to udev-trigger-early to
make sure graphics driver modules are loaded, so cirrus module
would start in VM.

BUG=chromium:611221,chrome-os-partner:53188,chrome-os-partner:53059
TEST=boot chromebook and look for frecon segfault in dmesg, also reboot
multiple times and observe no hangs with white screen on boot

Change-Id: Id0db9cde348034a01c0459e85360aa2c879486c9
Signed-off-by: Dominik Behr <dbehr@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/344418
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>