power: suspend_stress_test: Add flag to record dmesg(s) for failed iterations
Add a flag that records (only) the relevant part of dmesg for a failed
iteration. For instance, if the command used was:

"suspend_stress_test -c 10000 --noerrors_fatal --record_dmesg_dir=./logs"

and there were a total of 4 failured iterations, then the logs directory
might contain (as an example):
./dmesg_24
./dmesg_5423
./dmesg_7659
./dmesg_9211
The numbers in the file names indicate which iteration failed, and each
file contains only the relevent part of dmesg for that iteration (i.e.
logs start with the suspend cycle for that iteration)

I was testing a broken HW recently where suspend was failing on random
devices. I wanted to run full 10000 iteration to get a feel of how many
failures I see and which are the devices that actually fail. I realized
that at the end it only tells me the number of failures. It was very
difficult (or painful) to actually dig down the logs for dmesg for those
failed iterations among the thousands of iterations.

I modified the script to add this feature (that records dmesg of failed
iterations) and hopefully this might be helpful to others.

Note that the flag is off by default since the file system is read-only
by default.

BUG=chromium:614498
TEST=Tested on a HW that fails suspend_stress_test and verified that
     the test generates logs for failed iterations properly

Change-Id: Ie17a030b9b142dbc8f13f4d77ca8fd0fc4950e54
Signed-off-by: Rajat Jain <rajatja@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/347095
Reviewed-by: Dan Erat <derat@chromium.org>