shill: Add new histograms for time spent in suspend
When we resume from suspend, the wifi connection status
can be one of 4 scenarios as described by the
WiFiConnectionStatusOnWake metric. For each of these
scenarios, we would like to track the time spent in suspended
state. We consider it normal that the wifi disconnects after
a resume after a long time in suspend, but not after a short
time. This metric will help us remove the non-problematic
disconnect cases from the total number of disconnects, which
are reported in Network.Shill.WiFi.Disconnect.

BUG=chromium:614790
TEST=shill unit-test suite, ran suspend_stress_test on cyan

Change-Id: I2ed42a375c03538d8103548c9679319e78aba777
Reviewed-on: https://chromium-review.googlesource.com/347400
Commit-Ready: Kirtika Ruchandani <kirtika@google.com>
Tested-by: Kirtika Ruchandani <kirtika@google.com>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>