crash: before collecting a crash run filter helper
With this change before collecting a crash crash reporter will run
/opt/google/crash-reporter/filter if it exists and is executable.
PID will be passed as the only parameter. The crash will be collected
only if the filter returns 0.

This will be used to implement additional filtering in boards that
allow running arbitrary user binaries crashes from which should not
be collected.

BUG=b:28930938
TEST=Manually verified that crash-collecter runs the filter and does nothing if the filter returns non-zero status.

Change-Id: I660e7b6e8a60ee76545076338315cbd3b8d3a6d0
Reviewed-on: https://chromium-review.googlesource.com/347481
Commit-Ready: Andrey Ulanov <andreyu@google.com>
Tested-by: Andrey Ulanov <andreyu@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>