crosh: network_diag: remove excess escaping of asterisk
When creating the "network_diagnostics...txt" files, we don't need to
esapce an asterisk ("*") when it's quoted in a shell script. Perhaps the
original author was mixing up the escaping rules for regexes with the
rules for shell scripts? (I wouldn't blame 'em!)

This means instead of seeing diagnostic files named:

  network_diagnostics_\2016-05-25.19:07:50.txt

we'll now see:

  network_diagnostics_2016-05-25.19:07:50.txt

Amusingly, this bug also trips across the following ancient bug, which
means that the slash-named files can't be opened in Chrome (e.g., by
double clicking in the Files app):

https://bugs.chromium.org/p/chromium/issues/detail?id=45051

I guess I need a real file browser/text editor for that :)

BUG=none
TEST=run `network_diag` in shell and see what file(s) pop up

Change-Id: Id18052b512cb9688efbaef3b52c9e852fa17e541
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/347444
Reviewed-by: Jay Lee <jayhlee@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>