init/install: enable ext4 crypto option on stateful partition.
Allow using e4 crypto tool on stateful partition.

- When installing a new machine, if the kernel support ext4 crypto,
  format ext4 stateful partition with encryption support.
- At boot, if the kernel support ext4 crypto, but stateful fs does not
  have encryption support, enable it via tune2fs.

Note that until cryptohome is enabled to use ext4 crypto, this feature
will not be used.
Also, once a directory is encrypted, we should not use a kernel that
does not support ext4 crypto: it will be unable to access encrypted
directores properly (if \0 happens to be in the encrypted name).

TEST=On veyron_speedy: reboot machine with ext4 crypto enabled kernel,
check that 'encrypt' was added to the stateful filesystem feature.
Run chromeos-install, check the stateful filesystem was formated
properly.
Reinstall samus machine, check sda1 was set with encrypt option.
BUG=chromium:590820

Change-Id: I96254c200d9a28249aa0c7ff14f1a16adc132599
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/348380
Reviewed-by: Mike Frysinger <vapier@chromium.org>