dbus: power_manager: Add TABLET_MODE_{ON|OFF} protobuf definitions
This is part of the patch set that hooks up tablet mode switch
to Chrome. If the tablet mode switch is on then Chrome can know
for sure that it needs to enable maximize mode (tablet mode).
The helps Chrome to decide whether to enable tablet mode in some
corner cases where the accelerometer readings are not stable.

BUG=chrome-os-partner:49114
TEST=Build power_manager and chrome and deploy to my DUT. Make
     sure that:
     - The device enters maximize mode when tablet mode switch
       is on.
     - The device leaves maximize mode when tablet mode switch
       is off and lid angle is less than kExitMaximizeModeAngle.

Change-Id: Ia40584e5bc11d1d5d112e0684a47f8f133df34b1
Signed-off-by: Ricky Liang <jcliang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/347529
Reviewed-by: Dan Erat <derat@chromium.org>