libcontainer: Better error/resource management
Some error codes were not being propagated and some resources were not
being correctly torn down. This caused failure in one of the mount paths
to go down an error route, but the return code would not be updated,
causing the caller to believe it was successful, but the object was now
in a partially torn down state. This change fixes that and also fixes a
few additional leaks / use-after-free errors.

BUG=b:29101668
TEST=cheets_SettingsBridge, when in an intentionally bad state, fails
     correctly and reports an error instead of crashing.

Change-Id: I4059ea9e160031841f471a93dd56cdfb97eafa84
Reviewed-on: https://chromium-review.googlesource.com/349362
Commit-Ready: Luis Hector Chavez <lhchavez@google.com>
Tested-by: Luis Hector Chavez <lhchavez@google.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>