crash: Fix collection failure for ARC crash bursts
Since the timestamp in the filename of a crash report is only precise to
the second, collection failed for processes crashing more than once in
the span of a second. This CL provides a workaround, as well as minor
logging tweaks to help diagnose collection failures in the future.

BUG=b:29023301
TEST=Use Dev Tools to spam WTFs, and verify successful collection.

Change-Id: Ifb88b34e83567e523cdb22b1053d0a38ec7facb9
Reviewed-on: https://chromium-review.googlesource.com/350382
Commit-Ready: Dominik Laskowski <domlaskowski@google.com>
Tested-by: Dominik Laskowski <domlaskowski@google.com>
Reviewed-by: Luis Hector Chavez <lhchavez@google.com>