cryptohome: Remove sending ecryptfs key from crypto inferface.
Crypto is just passing through the key blob to the platform layer.
Have directly mount sends the keys to platform.

Fix an error in PushVaultKey where the error code was always true.

Change the interface to return true/false, instead of long, to be coherent
with the rest of the platform interface.

BUG=chromium:347322
TEST=Unitest, login_Cryptohome, platform_CryptohomeFio tests.

Change-Id: Ie423bbe66484eb87267c060b8e71daabe74ecc59
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/353355