quipper: Read buildids of non-kernel DSOs out of /proc/<tid>/root
Reading from proc/<tid>/root accounts for the thread's mnt namespace and
chroot, without the complication of entering a container.

If not found (likely the thread has exited), also tries reading from
/proc/<pid>/root, and finally also tries reading it directly.

If the DSO mapping is from a MMAP2 record, then the file's device major,
minor, and inode numbers must match those in the MMAP2 record. If it is not
identified as the same file, the DSO is not examined for a buildid. If
the device/inode info are not available, then the file is tried anyway.

TEST=Added extensive unit tests.
BUG=none

Change-Id: I24262f7494cf222ad8e504390bf8155227319c52
Reviewed-on: https://chromium-review.googlesource.com/354247
Commit-Ready: David Sharp <dhsharp@chromium.org>
Tested-by: David Sharp <dhsharp@chromium.org>
Reviewed-by: Chong Jiang <chongjiang@chromium.org>