power: Refactor KeyboardBacklightController.
Consolidate the logic for updating the keyboard backlight
brightness into the existing UpdateState() function.

No user-visible changes intended (apart from turning the
backlight on more quickly when fullscreen video stops), but
this paves the way for a future change to keep the backlight
on for a short period after hovering or user activity during
fullscreen video to reduce jank when hover detection is
noisy.

BUG=chromium:623404
TEST=existing tests pass with after one minor change

Change-Id: If758688a331841e517faab4c5f81cb930604189b
Reviewed-on: https://chromium-review.googlesource.com/356850
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>