libcontainer: copy_minor: Do not fail when host device does not exist
It is useful to be able to specify -1 as minor number in runtime.json,
so that the minor number is copied from outside the container.

However, the device may not exist outside of the container, in that case,
just skip it and not create it inside the container.

BUG=chrome-os-partner:54825
TEST=HWTest on cyan-cheets, veyron_minnie-cheets

Change-Id: I7d4c17a91cd421765a1764bc3b9fcfa64152e0d1
Reviewed-on: https://chromium-review.googlesource.com/357701
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>