cryptohome: Simplify homedir unit test.
Some of the tests were written with mysterious numbers without
annotation that it's hard to tell if it's necessary or not.

use 0 instead of kFreeSpace - 1 (if having not enough space is
important).

Do not just have a copy of Cache deletion test in GCache deletion test.

TEST=P2_TEST_FILTER="FreeDiskSpaceTest.*" \
  cros_workon_make --test cryptohome --board ${BOARD}
BUG=None

Change-Id: I68548244eef894ace70f2eff77b140e586713830
Reviewed-on: https://chromium-review.googlesource.com/358331
Commit-Ready: Junichi Uekawa <uekawa@chromium.org>
Tested-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>