shill: Allow duplicate routes with different metrics
When a new route shows up, RouteMsgHandler() currently scans the list of
known routes for matches.  If a matching route with a different metric
is found, the old entry is replaced by the new entry.

But on IPv6 it is normal for the kernel to add duplicate default routes
with different metrics in response to RA messages received off the LAN.
So shill needs to track the duplicate routes so that it knows which
entries to have the kernel delete when the service sort order changes.

BUG=chromium:625229
TEST=`FEATURES=test emerge-link shill`

Change-Id: I22cd572d891e800b41d3ff9b7a9e441fd3d5e3cc
Reviewed-on: https://chromium-review.googlesource.com/358172
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@google.com>