libcontainer: Run containers in a cgroup namespace
Now that all of our supported kernels support cgroup namespaces, enable
them when running a container.  This will allow the container to modify
cgroup parameters for sub groups but still allows the system to control
overall limits.  The contained processes will not know that it is in a
cgroup namespace, it appears as the root inside the namespace.

BUG=b/29259708
TEST=boot android on cyan and check cgroup membership inside and outside
the container.

Change-Id: I83e93ea6135f2b3f4baa4816eacacd1e02f1986a
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/362350