libcontainer: Only configure device cgroup if root
The CAP_SYS_ADMIN capability is needed to write to
the device cgroup files, and CAP_MKNOD is required
to call mknod. These capabilities are restricted to
root, so this commit checks that the user is root
before attempting to modify the device cgroup.

BUG=chromium:625285
TEST=Run container as root and as non-root and verify
that the device cgroup modification is skipped in
the non-root case.

Change-Id: Id3b227023a92c09b025a2b2397b4f70ce90b3098
Reviewed-on: https://chromium-review.googlesource.com/361595
Commit-Ready: Keshav Santhanam <ksanthanam@google.com>
Tested-by: Keshav Santhanam <ksanthanam@google.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>