libcontainer: Add option for pre-mounted runfs path
This allows the user to specify a path for the directory
in which they want the container to run. It also allows
the user to specify the path to the pid file.

BUG=chromium:628373
TEST=Create a directory, bind mount the container rootfs
to the new directory, use container_config_premounted_runfs
and container_config_pid_file_path.

Change-Id: I540700e529bbba8184dc17c0516ecc7ae7a6f793
Reviewed-on: https://chromium-review.googlesource.com/360580
Commit-Ready: Keshav Santhanam <ksanthanam@google.com>
Tested-by: Keshav Santhanam <ksanthanam@google.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>