Make component registration atomic, and check version rollback.
This changes makes the component registration atomic, so that when
updating a component, a crash does not leave the component in a bad
state. This is accomplished by having generational folders of a
component and a symlink that is atomically updated to point at the most
recent.

BUG=chromium:588220
TEST=emerge imageloader and run the test client with the test data
CQ-DEPEND=CL:361410,CL:361584

Change-Id: I731eab542e095e3c5b43e194374cf891bf302426
Reviewed-on: https://chromium-review.googlesource.com/361560
Commit-Ready: Greg Kerr <kerrnel@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Ricky Zhou <rickyz@chromium.org>