libcontainer: Make bind+ro similar to minijail.
minijail does a special handling of MS_BIND with MS_RDONLY, add that
handling to libcontainer as well so that adding or removing
"mount_outside" won't lose MS_RDONLY flag.

First bind mount will be able to setcontext, the remount will not be
able to setcontext because it's read-only.

BUG=b:29233242
TEST=test_that --board=samus-cheets $IP cheets_FileSystemPermissions \
  cheets_SELinuxTest
TEST=strace -p (session_manager pid) -f
  and observe the series of mount commands.

Change-Id: I9cda49e3b5896ee2f6ea6db5e3a29e7b09e67d56
Reviewed-on: https://chromium-review.googlesource.com/362201
Commit-Ready: Junichi Uekawa <uekawa@chromium.org>
Tested-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>