libcontainer: Add ability to configure cgroup owner
The owner and parent directory of the created cgroups needs to be
configured so that the new cgroups aren't created at the root of the
cgroup hierarchy.  This is useful when running a container in a cgroup
namespace, as the new cgroup namespace will have a subdirectory appear
as the root cgroup inside the namespace.

BUG=none
TEST=start with dependent changes and see the parent cgroup appears as
the root cgroup inside the container.

Change-Id: I1293ca691e0390a922b85bffd6e1dd28fbc10338
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/362096
Reviewed-by: Dmitry Grinberg <dmitrygr@google.com>
Reviewed-by: Keshav Santhanam <ksanthanam@google.com>