libcontainer: Separate cgroup configuration from creation
This adds the ability to set the owner of the parent cgroups, this is
needed when entering a cgroup namespace as the subdirectories and their
files must be owned by the user that root maps to outside the container.
Move cgroup creation from where the config is created until later.  This
information isn't known by login manager when creating the container.
Moving it later allows it to be configured like other config options.

BUG=none
TEST=run with dependent changes and see that the cgroup is created with
the correct owner and at the correct place in the cgroup hierarchy.

Change-Id: Ia31fb7eebe562b36216ce547c9889bcffba7f257
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/362097
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>