login: Set cgroup parent directory for android containers
Android will now run in a new cgroup namespace.  This will allow it to
modify its own cgroups while not knowing that it is part of a larger
cgroup hierarchy.  Configure the external cgroup
"session_manager_containers" that will be the root of Android's cgroup
namespace.

BUG=b/29259708
TEST=Start android and check that it is in a new cgroup namespace rooted
at session_manager_containers.

Change-Id: Ic361c2cf0904d05b959fad51c53b2fd983b22bc9
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/362099
Reviewed-by: Dan Erat <derat@chromium.org>